refactored codes in student folder
