+++ nestjs-backend/src/students/dto/update.dto.ts
@@ -1,0 +1,5 @@
+/* eslint-disable prettier/prettier */
+import { PartialType } from '@nestjs/mapped-types';
+import { StudentDto } from './student.dto';
+
+export class UpdateUserDto extends PartialType(StudentDto) {}
