+++ nestjs-backend/src/students/student.module.ts
@@ -2,7 +2,7 @@
 import { Module } from "@nestjs/common";
 import { StudentController } from "./student.controller";
 import { StudentService } from "./student.service";
-import { Student } from "./student.model";
+import { Student } from "./entities/student.entity";
 import { TypeOrmModule } from "@nestjs/typeorm";
 
 @Module({
